// _db.types.ts
interface Story extends Document {
  _id: ObjectId;
  createdAt: string;
  updatedAt: string;
  initialized: boolean;
  title: string;
  story_title: string;
  public: boolean;
  summary: string;
  passage: string;
  story_art: {
    prompt: string;
    image_url: string;
  };
// ...
}